Version 5.1.0,  Revision 20373,  Hash 449bca633
MODIFIED *  Makefile config/mk/global/actions/compile_interfaces.mk config/mk/global/actions/compile_internal_libraries.mk config/mk/global/actions/compile_yambo.mk config/mk/global/actions/compile_ypp.mk config/mk/global/functions/cleaning.mk config/mk/global/functions/mk_exe.mk config/mk/global/functions/mk_lib.mk include/driver/version.h sbin/compilation/check_updated_locks.sh sbin/compilation/check_updated_sources.sh sbin/compilation/dependencies_element.sh sbin/compilation/helper.sh interface/INIT_QP.F modules/mod_memory.F ph-el/PHEL_Self_Energy.F ph-el/PHEL_k_nearest_table.F pol_function/X_Double_Grid_setup.F

NEW *  sbin/compilation/check_object.sh sbin/compilation/check_object_elemental.sh

Bugs:
- Fighting with the _nl compilation procedure.
- Fixed lack of wait command in dependencies search. This was leading to compilation instabilities.

Patch sent by:  Andrea Marini <andrea.marini@cnr.it>

+++ sbin/compilation/check_updated_locks.sh
@@ -74,41 +74,36 @@ fi
 #echo "SAVE" $save_dir 
 #echo "RESTORE" $restore_dir 
 #
-# Remove the lock
-#
-bdir=`basename $dir`
-#
 # Search for objects dependent of new/missing PROJECTS 
 # and move them in a dedicated folder
 #
 for lock in $unmatched
 do
- rm -f $dir/$lock.lock
- if grep -q "$lock" $dir/.objects; then
-  rm -f config/stamps_and_lists/${goal}.stamp 
-  rm -f config/stamps_and_lists/${target}.a.stamp 
-  #echo "LOCK $lock in .objects"
+ if [ "$VERB" == 1 ] ; then
+  echo "rm -f $dir/$lock.lock"
+ else
+  rm -f $dir/$lock.lock
  fi
+ #
+ if grep -q "$lock" $dir/.objects; then
+  if [ "$VERB" == 1 ] ; then
+   echo "rm -f config/stamps_and_lists/${goal}.stamp"
+   echo "rm -f config/stamps_and_lists/${target}.a.stamp "
+  else
+   rm -f config/stamps_and_lists/${goal}.stamp 
+   rm -f config/stamps_and_lists/${target}.a.stamp 
+  fi
+ fi
+ #
+ # Check for PJ specific modules and their childs
+ #
  if test -f "$dir/${lock}_project.dep"; then
   deps=`cat $dir/${lock}_project.dep`
-  for object in $deps
+  for obj in $deps
   do
-    if test -f "$dir/$object"; then
-     if ! test -d $save_dir; then mkdir -p $dir/$save_dir; fi
-     rm -f config/stamps_and_lists/${goal}.stamp 
-     rm -f config/stamps_and_lists/${target}.a.stamp 
-     mv $dir/$object $dir/$save_dir
-     #echo "OBJ $object -> $save_dir"
-    fi
+    operate="save_and_restore"
+    source ./sbin/compilation/check_object.sh
   done
  fi
 done
 #
-# Restore the files of the project (if there are)
-#
-if test -d "$dir/$restore_dir"; then
- for file in `ls $dir/$restore_dir`
- do
-  mv $dir/$restore_dir/$file $dir
- done
-fi
